Navbar context layout completed
