+++ src/App/App.scss
@@ -1,4 +1,5 @@
 .App {
   display: grid;
   grid-template-columns: 72px 1fr;
+  font-family: Helvetica Neue;
 }
