+++ src/components/Navbar/Navbar.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import '../../styles/variables';
 
 .custom-navbar {
   background-color: $gray-800;
