+++ src/components/NavbarContext/NavbarContext.scss
@@ -1,0 +1,46 @@
+@import '../../styles/variables';
+
+.navbar-context {
+  position: absolute;
+  top: 0;
+  left: 72px;
+  height: 100vh;
+  background-color: $gray-100;
+  width: 216px;
+  box-shadow: 2px 0px 4px rgba(0, 0, 0, 0.25);
+}
+
+.context-header {
+  margin-top: 28px;
+  padding-bottom: 22px;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 20px;
+  line-height: 150%;
+  color: $gray-600;
+  border-bottom: 1px solid rgba($gray-500, 0.75);
+
+  h4 {
+    margin-bottom: 0px;
+  }
+
+  svg {
+    margin-right: 15px;
+  }
+}
+
+.context-items {
+  margin-top: 2.625rem;
+  margin-left: 20px;
+  color: $gray-700;
+
+  font-family: Helvetica Neue;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 18px;
+  line-height: 150%;
+
+  .context-item {
+    margin-bottom: 22px;
+  }
+}
