+++ src/styles/_variables.scss
@@ -1,4 +1,6 @@
-$favColor: beige;
+/**********/
+/* COLORS */
+/**********/
 
 /* White */
 $white: #FFFFFF;
